Revert "move serialization into core so it can be used in core. avoids circular dependency"
This reverts commit f4d2648ac481a7000446e4c333d40b57e6310958.
