+++ serialization/build.gradle
@@ -19,11 +19,10 @@ plugins {
 
 dependencies {
   implementation (project(":core"))
-  implementation (project(":serialization"))
   implementation (project(":jvm-libs:extensions"))
-  implementation "org.rocksdb:rocksdbjni"
-  implementation "tech.pegasys.teku.internal:storage"
-  implementation "org.hyperledger.besu.internal:metrics-core"
-  implementation "org.hyperledger.besu:plugin-api"
+  implementation "io.tmio:tuweni-bytes"
+  implementation "org.hyperledger.besu.internal:rlp"
+  implementation "org.hyperledger.besu:besu-datatypes"
+  testImplementation (project(":core"))
   testImplementation (testFixtures(project(":core")))
 }
