+++ serialization/src/main/kotlin/maru/serialization/rlp/BeaconBlockBodySerializer.kt
