+++ serialization/src/main/kotlin/maru/serialization/rlp/BeaconStateSerializer.kt
