+++ serialization/src/test/kotlin/maru/serialization/rlp/BeaconStateSerializerTest.kt
@@ -13,38 +13,27 @@
    See the License for the specific language governing permissions and
    limitations under the License.
  */
-package maru.core.serialization.rlp
+package maru.serialization.rlp
 
 import kotlin.random.Random
 import kotlin.random.nextULong
-import maru.core.BeaconBlock
-import maru.core.BeaconBlockBody
 import maru.core.BeaconBlockHeader
+import maru.core.BeaconState
 import maru.core.HashUtil
-import maru.core.Seal
 import maru.core.Validator
-import maru.core.ext.DataGenerators.randomExecutionPayload
-import maru.serialization.rlp.BeaconBlockBodySerializer
-import maru.serialization.rlp.BeaconBlockHeaderSerializer
-import maru.serialization.rlp.BeaconBlockSerializer
-import maru.serialization.rlp.ExecutionPayloadSerializer
-import maru.serialization.rlp.SealSerializer
-import maru.serialization.rlp.ValidatorSerializer
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 
-class BeaconBlockSerializerTest {
+class BeaconStateSerializerTest {
+  private val validatorSerializer = ValidatorSerializer()
+  private val beaconBlockHeaderSerializer =
+    BeaconBlockHeaderSerializer(
+      validatorSerializer = validatorSerializer,
+    )
   private val serializer =
-    BeaconBlockSerializer(
-      beaconBlockHeaderSerializer =
-        BeaconBlockHeaderSerializer(
-          validatorSerializer = ValidatorSerializer(),
-        ),
-      beaconBlockBodySerializer =
-        BeaconBlockBodySerializer(
-          sealSerializer = SealSerializer(),
-          executionPayloadSerializer = ExecutionPayloadSerializer(),
-        ),
+    BeaconStateSerializer(
+      beaconBlockHeaderSerializer = beaconBlockHeaderSerializer,
+      validatorSerializer = validatorSerializer,
     )
 
   @Test
@@ -60,16 +49,11 @@ class BeaconBlockSerializerTest {
         bodyRoot = Random.nextBytes(32),
         HashUtil::headerOnChainHash,
       )
-    val beaconBlockBody =
-      BeaconBlockBody(
-        prevCommitSeals = buildList(3) { Seal(Random.nextBytes(96)) },
-        commitSeals = buildList(3) { Seal(Random.nextBytes(96)) },
-        executionPayload = randomExecutionPayload(),
-      )
     val testValue =
-      BeaconBlock(
-        beaconBlockHeader = beaconBLockHeader,
-        beaconBlockBody = beaconBlockBody,
+      BeaconState(
+        latestBeaconBlockHeader = beaconBLockHeader,
+        latestBeaconBlockRoot = Random.nextBytes(32),
+        validators = buildSet(3) { Validator(Random.nextBytes(128)) },
       )
     val serializedData = serializer.serialize(testValue)
     val deserializedValue = serializer.deserialize(serializedData)
