+++ settings.gradle
@@ -11,3 +11,4 @@ include ':e2e'
 include 'jvm-libs'
 include 'jvm-libs:extensions'
 include 'storage'
+include 'serialization'
